requirements.txt
pandas 
streamlit==1.35.0
streamlit-option-menu
PyYaml
typer
matplotlib